eh: fix -Wshift-count-overflow in eh_setup_dcmd() (v6.10 preparation)
Upstream commit [1] changed the __ffs() return type from unsigned long
to unsigned int causing a -Wshift-count-overflow error:

drivers/soc/google/eh/eh_main.c:1133:44: error: shift count >= width of type [-Werror,-Wshift-count-overflow]
 1133 |         src_data = src_data = (__ffs(alignment) - 5) << EH_DCMD_BUF_SIZE_SHIFT;
      |                                                      ^  ~~~~~~~~~~~~~~~~~~~~~~
1 error generated.

Fix it by casting the value to be shifted to u64. As a note for the
reader, the bitfield manipulations can later be done using FIELD_PREP(),
which will make the code easier to read and hide this explicit cast.
Not in the scope of this surgical fix though.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=9c313ccdfc079f71f16c9b3d3a6c6d60996b9367 [1]
	("bitops: Change function return types from long to int")
Test: TH
Bug: 355381571
(cherry picked from commit 04f830bf59673e2823c12658e3b1dee88884d4c2)

+++ drivers/soc/google/eh/eh_main.c
@@ -1131,7 +1131,7 @@ static void eh_setup_dcmd(struct eh_device *eh_dev, unsigned int index,
 		       eh_dev->regs + EH_REG_DCMD_RES(index));
 #endif
 
-	src_data = (__ffs(alignment) - 5) << EH_DCMD_BUF_SIZE_SHIFT;
+	src_data = ((u64)(__ffs(alignment) - 5)) << EH_DCMD_BUF_SIZE_SHIFT;
 	src_data |= src_paddr;
 
 	/*
